I broke app yesterday and I repair it now, yeeeah

+++ src/main/java/data/abstracts/Thing.java
@@ -4,7 +4,7 @@ public abstract class Thing implements Item {
     private long id;
     private String title;
     private double price;
-    public Thing(int id, String title, double price){
+    public Thing(long id, String title, double price){
         this.id = id;
         this.title = title;
         this.price = price;
