+++ src/main/java/data/things/Car.java
@@ -13,7 +13,7 @@ public class Car extends Thing{
         return horsePower;
     }
 
-    public Car(int id, String title, double price, int horsePower) {
+    public Car(long id, String title, double price, int horsePower) {
         super(id, title, price);
         this.horsePower = horsePower;
     }
